add : 컴포넌트화 구현
