+++ src/layout/index.js
@@ -1,0 +1,2 @@
+export * from "./footer";
+export * from "./header";
