+++ src/pages/index.js
@@ -1,0 +1,4 @@
+export * from "./error-page";
+export * from "./home-page";
+export * from "./login-page";
+export * from "./profile-page";
