+++ src/route/index.js
@@ -1,0 +1,106 @@
+import { getUser, removeUser, setUser } from "../utils/storage";
+import { LoginPage, MainPage, ErrorPage, ProfilePage } from "../pages";
+
+const Router = () => {
+  const path = window.location.pathname;
+  const user = getUser();
+
+  if (user && path === "/login") {
+    history.pushState({}, "", "/");
+    Router();
+    return;
+  }
+
+  if (path === "/") {
+    document.getElementById("root").innerHTML = MainPage(user);
+  } else if (path === "/profile") {
+    if (user) {
+      document.getElementById("root").innerHTML = ProfilePage(user);
+    } else {
+      alert("로그인이 필요합니다");
+    }
+  } else if (path === "/login") {
+    document.getElementById("root").innerHTML = LoginPage();
+  } else {
+    document.getElementById("root").innerHTML = ErrorPage();
+  }
+  goToHome();
+  updateProfile();
+};
+
+const goToHome = () => {
+  const loginForm = document.getElementById("login-form");
+  if (loginForm) {
+    loginForm.addEventListener("submit", (e) => {
+      e.preventDefault();
+      const form = e.currentTarget;
+      const username = form.querySelector("#username").value.trim();
+
+      if (username.length) {
+        localStorage.setItem(
+          "user",
+          JSON.stringify({ username, email: "", bio: "" }),
+        );
+        history.pushState({}, "", "/");
+        Router();
+      } else {
+        alert("이메일과 비밀번호를 모두 입력해주세요.");
+      }
+    });
+  }
+};
+
+const updateProfile = () => {
+  const profileForm = document.getElementById("profile-form");
+  if (profileForm) {
+    profileForm.addEventListener("submit", (e) => {
+      e.preventDefault();
+      const username = profileForm.querySelector("#username").value.trim();
+      const email = profileForm.querySelector("#email").value.trim();
+      const bio = profileForm.querySelector("#bio").value.trim();
+      const user = JSON.parse(localStorage.getItem("user"));
+      const updateUser = { ...user, username, email, bio };
+
+      setUser(updateUser);
+      alert("프로필이 수정되었습니다");
+      Router();
+    });
+  }
+};
+
+const goToLogin = () => {
+  removeUser();
+  history.pushState({}, "", "/");
+  Router();
+};
+
+const initRouterEvent = () => {
+  document.addEventListener("click", (e) => {
+    const target = e.target.closest("a");
+    if (target && target.id === "logout") {
+      e.preventDefault();
+      goToLogin();
+    }
+
+    if (
+      target &&
+      target.getAttribute("href") &&
+      target.getAttribute("href").startsWith("/")
+    ) {
+      e.preventDefault();
+      const href = target.getAttribute("href");
+      history.pushState({}, "", href);
+      Router();
+    }
+  });
+};
+
+export const App = {
+  init: () => {
+    Router();
+    window.addEventListener("popstate", () => {
+      Router();
+    });
+    initRouterEvent();
+  },
+};
