+++ src/utils/storage.js
@@ -1,0 +1,11 @@
+export const setUser = (user) => {
+  localStorage.setItem("user", JSON.stringify(user));
+};
+
+export const getUser = () => {
+  return JSON.parse(localStorage.getItem("user"));
+};
+
+export const removeUser = () => {
+  localStorage.removeItem("user");
+};
